Fix: Refactor controller folder
