+++ Backend/Gym-tracker/Gym-tracker/Controllers/User/AuthenticationController.cs
@@ -1,9 +1,8 @@
-﻿
-using Members.API.Dtos;
+﻿using Members.API.Dtos;
 using Members.API.Public;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Gym_tracker.Controllers
+namespace Gym_tracker.Controllers.User
 {
     [Route("api/auth")]
