--- Backend/Gym-tracker/Gym-tracker/Controllers/UserController.cs
@@ -1,23 +1,0 @@
-﻿using BuildingBlocks.Core.UseCases;
-using Members.API.Dtos;
-using Members.API.Public;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Gym_tracker.Controllers
-{
-    [Route("api/user")]
-    public class UserController : BaseApiController
-    {
-        private readonly IUserService _userService;
-        public UserController(IUserService userService)
-        {
-            _userService = userService;
-        }
-        [HttpGet]
-        public ActionResult<PagedResult<UserDto>> GetAll([FromQuery] int page, [FromQuery] int pageSize)
-        {
-            var result = _userService.GetPaged(page, pageSize);
-            return CreateResponse(result);
-        }
-    }
-}
