+++ Backend/Gym-tracker/Gym-tracker/Controllers/Workout/WorkoutController.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using Training.API.Dtos;
 using Training.API.Public;
 
-namespace Gym_tracker.Controllers
+namespace Gym_tracker.Controllers.Workout
 {
     [Route("api/workout")]
     [Authorize]
